+++ packages/wildness_ui_golden_toolkit/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1
+chantes on testColumnComponent and testDeviceComponent:
+- removed gestureBuilder parameter as required
+- added groupName parameter
 ## 0.1.0
-
-- First version
 - First version